moving search into engine
