+++ lua/grug-far.lua
@@ -4,6 +4,7 @@ local farBuffer = require('grug-far/farBuffer')
 local history = require('grug-far/history')
 local utils = require('grug-far/utils')
 local close = require('grug-far/actions/close')
+local engine = require('grug-far/engine')
 
 local M = {}
 
@@ -125,6 +126,7 @@ local contextCount = 0
 ---@field state GrugFarState
 ---@field prevWin? integer
 ---@field actions GrugFarAction[]
+---@field engine GrugFarEngine
 
 --- generate instance specific context
 ---@param options GrugFarOptions
@@ -134,6 +136,7 @@ local function createContext(options)
   return {
     count = contextCount,
     options = options,
+    engine = engine.getEngine(options.engine),
     namespace = vim.api.nvim_create_namespace('grug-far-namespace'),
     locationsNamespace = vim.api.nvim_create_namespace(''),
     historyHlNamespace = vim.api.nvim_create_namespace(''),
